(#33) Move HPC helpers from runs/ to pycsa/scheduling.py
- New module `pycsa/scheduling.py` hosts `estimate_cell_memory_gb` and `group_cells_by_memory` (pure numpy; no Dask at import time). Behavior is byte-identical to the originals in `runs/icon_etopo_global.py`.
- `runs/icon_etopo_global.py` re-imports from `pycsa.scheduling` and drops the local defs (-144 lines).
- `tests/test_dynamic_memory.py` drops its `sys.path.insert("/home/ray/git-projects/spec_appx/runs")` hack and imports cleanly from `pycsa.scheduling`. Test passes in 1.2 s without pulling Dask.
- Reproducibility suite still green (3 passed in 14.4 s).
- changelog.d/33.added.rst (towncrier).

First of four Phase B refactors gated by the reproducibility suite (#31). Sequence: #33 → #34 → #35 → #36.
